requirements.txt
deepgram-sdk
deepgram-captions
httpx